# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "D-Care Console API",
    "description": "치과 아웃바운드 관리 시스템 API\n\n## 인증\n모든 V2 API는 JWT Bearer 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 포함하세요.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "현재 서버" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalCount": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "CallLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "phone": { "type": "string" },
          "callType": { "type": "string", "enum": ["inbound", "outbound"] },
          "duration": { "type": "integer" },
          "callTime": { "type": "string", "format": "date-time" },
          "patientId": { "type": "string" },
          "patientName": { "type": "string" },
          "classification": { "type": "string", "enum": ["환자", "거래처", "스팸", "기타"] },
          "summary": { "type": "string" },
          "interest": { "type": "string" },
          "temperature": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed"] },
          "callbackType": { "type": "string", "enum": ["callback", "recall", "thanks"] }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v2/call-logs": {
      "get": {
        "tags": ["Call Logs"],
        "summary": "통화 기록 목록 조회",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "direction", "in": "query", "schema": { "type": "string", "enum": ["inbound", "outbound"] } },
          { "name": "classification", "in": "query", "schema": { "type": "string", "enum": ["환자", "거래처", "스팸", "기타"] } },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "전화번호 검색" },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "startedAt" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": {
            "description": "통화 기록 목록 + 분류 통계",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "callLogs": { "type": "array", "items": { "$ref": "#/components/schemas/CallLog" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" },
                    "stats": { "type": "object", "description": "분류별 통화 건수" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Call Logs"],
        "summary": "통화 기록 AI 분석 결과 수정 / 환자 연결",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["callLogId"],
                "properties": {
                  "callLogId": { "type": "string" },
                  "classification": { "type": "string", "enum": ["환자", "거래처", "스팸", "기타"] },
                  "patientId": { "type": "string", "nullable": true, "description": "환자 연결 (null이면 연결 해제)" },
                  "patientName": { "type": "string" },
                  "interest": { "type": "string" },
                  "temperature": { "type": "string" },
                  "summary": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "수정 성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "callLog": { "$ref": "#/components/schemas/CallLog" } } } } } }
        }
      }
    },
...
  },
...
}